add health bar

+++ Forestious_3D/Assets/Scripts/CharacterStats.cs
@@ -2,6 +2,6 @@ using UnityEngine;
 
 public class CharacterStats : MonoBehaviour
 {
-    public int health = 3;
+    public float health = 3;
     public float moveSpeed = 7.5f;
 }
